Modularized app into document fetchers and extractors separate from main app flow
